# memory-hierachy 2 prework solution