--- a/src/classes/intro-systems/memory-hierarchy-2/prework/main.go
+++ b/src/classes/intro-systems/memory-hierarchy-2/prework/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "fmt"
+
+func main() {
+	fmt.Println("This is just a stub for now.")
+}