--- a/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/bench.sh
+++ b/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/bench.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+cd "$(dirname "${BASH_SOURCE[0]}")"
+set -euxo pipefail
+
+go test -bench=.