--- a/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/metrics.go
+++ b/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/metrics.go
@@ -11,49 +11,54 @@ import (
 type UserId int
 type UserMap map[UserId]*User
 
-type User struct {
-	id       UserId
-	age      int
-	payments []int
+type Users struct {
+	userMap UserMap
+
+	allAges     []int
+	allPayments []int
 }
 
-func AverageAge(users UserMap) float64 {
+type User struct {
+	id UserId
+
+	ageIndex       int
+	paymentIndexes []int
+}
+
+func AverageAge(users Users) float64 {
 	average, count := 0.0, 0.0
-	for _, u := range users {
+	for _, age := range users.allAges {
 		count += 1
-		average += (float64(u.age) - average) / count
+		average += (float64(age) - average) / count
 	}
 	return average
 }
 
-func AveragePaymentAmount(users UserMap) float64 {
+func AveragePaymentAmount(users Users) float64 {
 	average, count := 0.0, 0.0
-	for _, u := range users {
-		for _, p := range u.payments {
-			count += 1
-			amount := float64(p)
-			average += (amount - average) / count
-		}
+	for _, p := range users.allPayments {
+		count += 1
+		amount := float64(p)
+		average += (amount - average) / count
 	}
+
 	return average / 100
 }
 
 // Compute the standard deviation of payment amounts
-func StdDevPaymentAmount(users UserMap) float64 {
+func StdDevPaymentAmount(users Users) float64 {
 	mean := AveragePaymentAmount(users)
 	squaredDiffs, count := 0.0, 0.0
-	for _, u := range users {
-		for _, p := range u.payments {
-			count += 1
-			amount := float64(p / 100)
-			diff := amount - mean
-			squaredDiffs += diff * diff
-		}
+	for _, p := range users.allPayments {
+		count += 1
+		amount := float64(p / 100)
+		diff := amount - mean
+		squaredDiffs += diff * diff
 	}
 	return math.Sqrt(squaredDiffs / count)
 }
 
-func LoadData() UserMap {
+func LoadData() Users {
 	f, err := os.Open("users.csv")
 	if err != nil {
 		log.Fatalln("Unable to read users.csv", err)
@@ -64,11 +69,20 @@ func LoadData() UserMap {
 		log.Fatalln("Unable to parse users.csv as csv", err)
 	}
 
-	users := make(UserMap, len(userLines))
+	users := Users{
+		userMap: make(UserMap, len(userLines)),
+	}
+
 	for _, line := range userLines {
 		id, _ := strconv.Atoi(line[0])
+
 		age, _ := strconv.Atoi(line[2])
-		users[UserId(id)] = &User{id: UserId(id), age: age}
+		users.allAges = append(users.allAges, age)
+
+		users.userMap[UserId(id)] = &User{
+			id:       UserId(id),
+			ageIndex: len(users.allAges) - 1,
+		}
 	}
 
 	f, err = os.Open("payments.csv")
@@ -84,7 +98,12 @@ func LoadData() UserMap {
 	for _, line := range paymentLines {
 		userID, _ := strconv.Atoi(line[2])
 		paymentCents, _ := strconv.Atoi(line[0])
-		users[UserId(userID)].payments = append(users[UserId(userID)].payments, paymentCents)
+
+		users.allPayments = append(users.allPayments, paymentCents)
+		id := UserId(userID)
+
+		user := users.userMap[id]
+		user.paymentIndexes = append(user.paymentIndexes, len(users.allPayments)-1)
 	}
 
 	return users