--- a/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/metrics_test.go
+++ b/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/metrics_test.go
@@ -41,5 +41,4 @@ func BenchmarkMetrics(b *testing.B) {
 			b.Fatalf("Expected standard deviation to be around %.2f, not %.3f", expected, actual)
 		}
 	})
-
 }