--- a/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/start.txt
+++ b/src/classes/intro-systems/memory-hierarchy-2/prework/metrics/start.txt
@@ -1,0 +1,10 @@
+geoffrey@geoffrey-server ~/d/c/s/c/i/m/p/metrics (mem-golang)> ./bench.sh
++ go test -bench=.
+goos: linux
+goarch: amd64
+pkg: github.com/ggilmore/csi/src/classes/intro-systems/memory-hierarchy-2/prework/metrics
+cpu: Intel(R) Core(TM) i7-7700K CPU @ 4.20GHz
+BenchmarkMetrics/Average_age-8               475           2549404 ns/op
+BenchmarkMetrics/Average_payment-8            56          20336694 ns/op
+BenchmarkMetrics/Payment_stddev-8             28          39649129 ns/op
+PASS