(ignoring any optimization like loop unrolling, etc.)

| function               | before (`ns/op`) | after ( `ns/op`) | speedup |
|------------------------|------------------|------------------|---------|
| `AverageAge`           | `2549404`        | `488578`         | ~5.22x  |
| `AveragePaymentAmount` | `20336694`       | `4912768`        | ~4.14x  |
| `StdDevPaymentAmount`  | `39649129`       | `6227807`        | ~6.37x  |